committee save,draft Save, get draft, get Data

+++ student-welfare/src/main/java/in/cdac/university/studentWelfare/bean/SavitbpSchemeApplTrackermst.java
@@ -43,7 +43,7 @@ public class SavitbpSchemeApplTrackermst {
 	
 	private Integer unumIsvalid;
 	
-	private Integer unumEntryUid;
+	private Long unumEntryUid;
 	
 	private Date udtEntryDate;
